+++ spec/stat_tracker_spec.rb
@@ -74,38 +74,6 @@ RSpec.describe StatTracker do
     expect(@stat_tracker.average_goals_by_season).to eq({"20122013"=>3.5, "20142015"=>4.5, "20152016"=>4.0})
   end
 
-  it '#count_of_teams' do
-    expect(@stat_tracker.count_of_teams).to eq(32)
-  end
-
-  it '#average_goals' do
-    expect(@stat_tracker.average_goals).to eq({"16"=>1.0, "17"=>2.0, "24"=>2.0, "25"=>2.0, "26"=>3.0, "8"=>2.5, "9"=>1.0})
-  end
-
-  it '#best_offense' do
-    expect(@stat_tracker.best_offense).to eq('FC Cincinnati')
-  end
-
-  it '#worst_offense' do
-    expect(@stat_tracker.worst_offense). to eq('New England Revolution')
-  end
-
-  it '#highest_scoring_visitor' do
-    expect(@stat_tracker.highest_scoring_visitor).to eq('FC Cincinnati')
-  end
-
-  it '#highest_scoring_home_team' do
-    expect(@stat_tracker.highest_scoring_home_team).to eq('New York Red Bulls')
-  end
-
-  it '#lowest_scoring_visitor' do
-    expect(@stat_tracker.lowest_scoring_visitor).to eq('New England Revolution')
-  end
-
-  it '#lowest_scoring_home_team' do
-    expect(@stat_tracker.lowest_scoring_home_team).to eq('New York City FC')
-  end
-
   it '#most_goals_scored' do
     expect(@stat_tracker.most_goals_scored('8')).to eq(3)
   end
